requirements.txt
together==1.2.8
voyageai==0.2.3
langchain==0.3.1
langchain-community==0.3.0
langchain-core==0.3.6
langchain-openai==0.2.1
langchain-text-splitters==0.3.0
langgraph==0.2.16
faiss-cpu==1.8.0
pypdf==4.3.1
python-dotenv==1.0.1
streamlit==1.38.0
gradio==4.44.0
numpy==1.26.4
pathlib2==2.3.7
sentence-transformers==3.0.1
unstructured==0.15.12
sentence-transformers==3.3.1
unstructured[pdf]==0.16.10
